backend/requirements.dev.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
black==26.1.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.133.0
fonttools==4.61.1
h11==0.16.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
mypy_extensions==1.1.0
numpy==2.4.2
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-multipart==0.0.22
pytokens==0.4.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
